requirements.txt
absl-py==0.13.0
atari-py==0.2.9
autonomous-learning-library==0.6.2
box2d-py==2.3.8
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
future==0.18.2
google-auth==1.32.1
google-auth-oauthlib==0.4.4
grpcio==1.38.1
gym==0.18.3
gym-super-mario-bros==7.3.2
idna==2.10
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
nes-py==8.1.6
numpy==1.21.0
oauthlib==3.1.1
opencv-python==3.4.14.53
pandas==1.2.5
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.1.7
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.0
six==1.16.0
stable-baselines3==1.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
